requirements.txt
flask==3.0.3
flask-cors==4.0.0
groq==0.9.0
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.4.1
gunicorn==22.0.0
Pillow
requests
